restructure repo with services
