+++ src/schemas/strapi.py
