+++ src/services/sert.py
@@ -7,12 +7,12 @@ from fastapi import HTTPException
 from jinja2 import Template
 from vllm.sampling_params import SamplingParams
 
-from .dependencies.strapi import Strapi
-from .dependencies.supabase import SupabaseClient
-from .models.chat import EventType
-from .models.embedding import RetrievalInput, RetrievalStrategy
-from .models.summary import ChunkWithWeight, Summary
-from .pipelines.chat import chat_pipeline
+from ..dependencies.strapi import Strapi
+from ..dependencies.supabase import SupabaseClient
+from ..schemas.chat import EventType
+from ..schemas.embedding import RetrievalInput, RetrievalStrategy
+from ..schemas.summary import ChunkWithWeight, Summary
+from ..pipelines.chat import chat_pipeline
 
 with open("templates/sert.jinja2", "r", encoding="utf8") as file_:
     prompt_template = Template(file_.read())
