+++ src/services/transcript.py
@@ -3,7 +3,7 @@ from urllib import parse
 from fastapi import HTTPException
 from youtube_transcript_api import YouTubeTranscriptApi
 
-from .models.transcript import TranscriptInput, TranscriptResults
+from ..schemas.transcript import TranscriptInput, TranscriptResults
 
 
 async def transcript_generate(transcript_input: TranscriptInput) -> TranscriptResults:
